+++ src/HttpClient/HttpClientFactory.php
@@ -105,7 +105,7 @@ final class HttpClientFactory implements HttpClientFactoryInterface
             new HeaderSetPlugin(['User-Agent' => $this->sdkIdentifier . '/' . $this->sdkVersion]),
             new AuthenticationPlugin(new SentryAuthentication($options, $this->sdkIdentifier, $this->sdkVersion)),
             new RetryPlugin(['retries' => $options->getSendAttempts()]),
-            new ErrorPlugin(),
+            new ErrorPlugin(['only_server_exception' => true]),
         ];
 
         if ($options->isCompressionEnabled()) {